refactor(cli): extract context command family into cli/commands/context.ts
